Cleanup
Moved ReplicationClient under replica
ReplicationClient does not use dbms_handler (instead a callback is
provided)
CreateReplicationClient has been removed from Storage
Moved FrequentHeartbeat to own header
