+++ src/storage/v2/replication/replication_client.hpp
@@ -13,6 +13,8 @@
 
 #include "replication/config.hpp"
 #include "replication/epoch.hpp"
+#include "replication/messages.hpp"
+#include "replication/replication_client.hpp"
 #include "rpc/client.hpp"
 #include "storage/v2/durability/storage_global_operation.hpp"
 #include "storage/v2/id_types.hpp"
@@ -34,9 +36,6 @@
 #include <string>
 #include <variant>
 
-namespace memgraph::dbms {
-class DbmsHandler;
-}  // namespace memgraph::dbms
 namespace memgraph::storage {
 
 struct Delta;
@@ -48,7 +47,7 @@ class ReplicationStorageClient;
 // Handler used for transferring the current transaction.
 class ReplicaStream {
  public:
-  explicit ReplicaStream(Storage *storage, rpc::Client &rpc_client, const uint64_t current_seq_num);
+  explicit ReplicaStream(Storage *storage, rpc::Client &rpc_client, uint64_t current_seq_num);
 
   /// @throw rpc::RpcFailedException
   void AppendDelta(const Delta &delta, const Vertex &vertex, uint64_t final_commit_timestamp);
@@ -77,50 +76,13 @@ class ReplicaStream {
 template <typename F>
 concept InvocableWithStream = std::invocable<F, ReplicaStream &>;
 
-struct ReplicationClient {
-  explicit ReplicationClient(const memgraph::replication::ReplicationClientConfig &config);
-
-  ~ReplicationClient();
-  ReplicationClient(ReplicationClient const &) = delete;
-  ReplicationClient &operator=(ReplicationClient const &) = delete;
-  ReplicationClient(ReplicationClient &&) noexcept = delete;
-  ReplicationClient &operator=(ReplicationClient &&) noexcept = delete;
-
-  void Start(dbms::DbmsHandler *dbms_handler);
-  void StartFrequentCheck(dbms::DbmsHandler &dbms_handler);
-  void FrequentCheck(dbms::DbmsHandler *dbms_handler);
-
-  std::string name_;
-  communication::ClientContext rpc_context_;
-  rpc::Client rpc_client_;
-  std::chrono::seconds replica_check_frequency_;
-
-  memgraph::replication::ReplicationMode mode_{memgraph::replication::ReplicationMode::SYNC};
-  // This thread pool is used for background tasks so we don't
-  // block the main storage thread
-  // We use only 1 thread for 2 reasons:
-  //  - background tasks ALWAYS contain some kind of RPC communication.
-  //    We can't have multiple RPC communication from a same client
-  //    because that's not logically valid (e.g. you cannot send a snapshot
-  //    and WAL at a same time because WAL will arrive earlier and be applied
-  //    before the snapshot which is not correct)
-  //  - the implementation is simplified as we have a total control of what
-  //    this pool is executing. Also, we can simply queue multiple tasks
-  //    and be sure of the execution order.
-  //    Not having mulitple possible threads in the same client allows us
-  //    to ignore concurrency problems inside the client.
-  utils::ThreadPool thread_pool_{1};
-
-  utils::Scheduler replica_checker_;
-};
-
 class ReplicationStorageClient {
-  friend class CurrentWalHandler;
+  friend class InMemoryCurrentWalHandler;
   friend class ReplicaStream;
-  friend struct ReplicationClient;
+  friend struct ::memgraph::replication::ReplicationClient;
 
  public:
-  explicit ReplicationStorageClient(ReplicationClient &client);
+  explicit ReplicationStorageClient(::memgraph::replication::ReplicationClient &client);
 
   ReplicationStorageClient(ReplicationStorageClient const &) = delete;
   ReplicationStorageClient &operator=(ReplicationStorageClient const &) = delete;
@@ -164,16 +126,16 @@ class ReplicationStorageClient {
   // Return whether the transaction could be finalized on the replication client or not.
   [[nodiscard]] bool FinalizeTransactionReplication(Storage *storage);
 
- protected:
+  void TryCheckReplicaStateAsync(Storage *storage);  // TODO Move back to private
+ private:
   void RecoverReplica(uint64_t replica_commit, memgraph::storage::Storage *storage);
 
   void CheckReplicaState(Storage *storage);
   void LogRpcFailure();
-  void TryCheckReplicaStateAsync(Storage *storage);
   void TryCheckReplicaStateSync(Storage *storage);
   void FrequentCheck(Storage *storage);
 
-  ReplicationClient &client_;
+  ::memgraph::replication::ReplicationClient &client_;
   std::optional<ReplicaStream>
       replica_stream_;  // Currently active stream (nullopt if not in use), note: a single stream per rpc client
   mutable utils::Synchronized<replication::ReplicaState, utils::SpinLock> replica_state_{
