+++ src/storage/v2/replication/replication_storage_state.hpp
@@ -47,7 +47,7 @@ struct ReplicationStorageState {
 
   // Getters
   auto GetReplicaState(std::string_view name) const -> std::optional<replication::ReplicaState>;
-  auto ReplicasInfo(Storage *storage) const -> std::vector<ReplicaInfo>;
+  auto ReplicasInfo(const Storage *storage) const -> std::vector<ReplicaInfo>;
 
   // History
   void TrackLatestHistory();
